feat: Add ConfigModule and DatabaseModule to common library
This commit adds the ConfigModule and DatabaseModule to the common library in order to provide configuration management and database connectivity functionality. The ConfigModule is responsible for loading and validating environment variables, while the DatabaseModule sets up the connection to the MongoDB database using the Mongoose module. These modules will be used by other components in the application.

Refs: #123